+++ Editor/Plugins/NotInitializePluginAttribute.cs
@@ -1,5 +1,5 @@
 namespace UnityEditor.DeviceSimulation
 {
     [System.AttributeUsage(System.AttributeTargets.Class)]
-    internal class NotInitializePluginAttribute : System.Attribute {}
+    internal class NotInitializePluginAttribute : System.Attribute { }
 }