+++ Tests/Editor/Input/InputProviderMouseInput.cs
@@ -14,19 +14,11 @@ namespace Unity.DeviceSimulator.Editor.Tests.Input
         private TouchEventManipulator m_InputProvider;
         private ScreenSimulation m_ScreenSimulation;
 
-        [UnitySetUp]
-        public IEnumerator SetUp()
-        {
-            yield return new EnterPlayMode();
-        }
-
-        [UnityTearDown]
-        public IEnumerator TearDown()
+        [TearDown]
+        public void TearDown()
         {
             m_InputProvider?.Dispose();
             m_ScreenSimulation?.Dispose();
-
-            yield return new ExitPlayMode();
         }
 
         private void SetUpSimulation(DeviceInfo deviceInfo)
@@ -37,6 +29,7 @@ namespace Unity.DeviceSimulator.Editor.Tests.Input
             m_InputProvider = new TouchEventManipulator(new UnityEditor.DeviceSimulation.DeviceSimulator());
             m_ScreenSimulation = new ScreenSimulation(m_DeviceInfo, new SimulationPlayerSettings());
             m_InputProvider.InitTouchInput(null, deviceInfo, m_ScreenSimulation);
+            m_ScreenSimulation.ApplyChanges();
         }
 
         internal class InputProviderTouchFromMouseCase
@@ -55,28 +48,36 @@ namespace Unity.DeviceSimulator.Editor.Tests.Input
             yield return new InputProviderTouchFromMouseCase
             {
                 deviceInfo = DeviceInfoLibrary.GetGalaxy10e(),
-                screenOrientation = ScreenOrientation.Portrait, resolutionWidth = 1080, resolutionHeight = 2280,
+                screenOrientation = ScreenOrientation.Portrait,
+                resolutionWidth = 1080,
+                resolutionHeight = 2280,
                 inputTouchPosition = new[] { new Vector2(1, 1), new Vector2(1000, 1500), new Vector2(100, 100) },
                 resultTouchPosition = new[] { new Vector2(1, 2279), new Vector2(1000, 780), new Vector2(100, 2180) }
             };
             yield return new InputProviderTouchFromMouseCase
             {
                 deviceInfo = DeviceInfoLibrary.GetGalaxy10e(),
-                screenOrientation = ScreenOrientation.LandscapeLeft, resolutionWidth = 2280, resolutionHeight = 1080,
+                screenOrientation = ScreenOrientation.LandscapeLeft,
+                resolutionWidth = 2280,
+                resolutionHeight = 1080,
                 inputTouchPosition = new[] { new Vector2(1, 1), new Vector2(1000, 333), new Vector2(99, 100) },
                 resultTouchPosition = new[] { new Vector2(1, 1), new Vector2(333, 1000), new Vector2(100, 99) }
             };
             yield return new InputProviderTouchFromMouseCase
             {
                 deviceInfo = DeviceInfoLibrary.GetGalaxy10e(),
-                screenOrientation = ScreenOrientation.PortraitUpsideDown, resolutionWidth = 1080, resolutionHeight = 2280,
+                screenOrientation = ScreenOrientation.PortraitUpsideDown,
+                resolutionWidth = 1080,
+                resolutionHeight = 2280,
                 inputTouchPosition = new[] { new Vector2(1, 8), new Vector2(1000, 1503), new Vector2(100, 100) },
                 resultTouchPosition = new[] { new Vector2(1079, 8), new Vector2(80, 1503), new Vector2(980, 100) }
             };
             yield return new InputProviderTouchFromMouseCase
             {
                 deviceInfo = DeviceInfoLibrary.GetGalaxy10e(),
-                screenOrientation = ScreenOrientation.LandscapeRight, resolutionWidth = 2280, resolutionHeight = 1080,
+                screenOrientation = ScreenOrientation.LandscapeRight,
+                resolutionWidth = 2280,
+                resolutionHeight = 1080,
                 inputTouchPosition = new[] { new Vector2(1, 1), new Vector2(1000, 333), new Vector2(100, 100) },
                 resultTouchPosition = new[] { new Vector2(2279, 1079), new Vector2(1947, 80), new Vector2(2180, 980) }
             };
@@ -85,21 +86,27 @@ namespace Unity.DeviceSimulator.Editor.Tests.Input
             yield return new InputProviderTouchFromMouseCase
             {
                 deviceInfo = DeviceInfoLibrary.GetGalaxy10e(),
-                screenOrientation = ScreenOrientation.Portrait, resolutionWidth = 2000, resolutionHeight = 4000,
+                screenOrientation = ScreenOrientation.Portrait,
+                resolutionWidth = 2000,
+                resolutionHeight = 4000,
                 inputTouchPosition = new[] { new Vector2(1, 1), new Vector2(1079, 2279), new Vector2(99, 699) },
                 resultTouchPosition = new[] { new Vector2(1.85f, 3998.24f), new Vector2(1998.15f, 1.75f), new Vector2(183.33f, 2773.68f) }
             };
             yield return new InputProviderTouchFromMouseCase
             {
                 deviceInfo = DeviceInfoLibrary.GetGalaxy10e(),
-                screenOrientation = ScreenOrientation.LandscapeLeft, resolutionWidth = 1140, resolutionHeight = 540,
+                screenOrientation = ScreenOrientation.LandscapeLeft,
+                resolutionWidth = 1140,
+                resolutionHeight = 540,
                 inputTouchPosition = new[] { new Vector2(1, 1), new Vector2(1000, 333), new Vector2(99, 100) },
                 resultTouchPosition = new[] { new Vector2(0.5f, 0.5f), new Vector2(166.5f, 500), new Vector2(50, 49.5f) }
             };
             yield return new InputProviderTouchFromMouseCase
             {
                 deviceInfo = DeviceInfoLibrary.GetGalaxy10e(),
-                screenOrientation = ScreenOrientation.PortraitUpsideDown, resolutionWidth = 2280, resolutionHeight = 1080,
+                screenOrientation = ScreenOrientation.PortraitUpsideDown,
+                resolutionWidth = 2280,
+                resolutionHeight = 1080,
                 inputTouchPosition = new[] { new Vector2(1, 8), new Vector2(1000, 1503), new Vector2(100, 100) },
                 resultTouchPosition = new[] { new Vector2(2277.88f, 3.79f), new Vector2(168.88f, 711.94f), new Vector2(2068.88f, 47.36f) }
             };
@@ -109,6 +116,8 @@ namespace Unity.DeviceSimulator.Editor.Tests.Input
         public IEnumerator LegacyInputTouchFromMouse(
             [ValueSource("InputProviderTouchFromMouseCases")] InputProviderTouchFromMouseCase testData)
         {
+            yield return new EnterPlayMode();
+
             Assert.AreEqual(3, testData.inputTouchPosition.Length);
             Assert.AreEqual(3, testData.resultTouchPosition.Length);
 
@@ -121,11 +130,13 @@ namespace Unity.DeviceSimulator.Editor.Tests.Input
 
             SetUpSimulation(testData.deviceInfo);
             Screen.orientation = testData.screenOrientation;
+            m_ScreenSimulation.ApplyChanges();
 
             if (m_ScreenSimulation.Width != testData.resolutionWidth || m_ScreenSimulation.Height != testData.resolutionHeight)
             {
                 Screen.SetResolution(testData.resolutionWidth, testData.resolutionHeight, FullScreenMode.FullScreenWindow);
             }
+            m_ScreenSimulation.ApplyChanges();
 
             yield return null;
 
@@ -164,6 +175,8 @@ namespace Unity.DeviceSimulator.Editor.Tests.Input
             yield return null;
 
             Assert.AreEqual(0, UnityEngine.Input.touchCount);
+
+            yield return new ExitPlayMode();
         }
     }
 }